app/src/main/java/kpn/lib/domain/AbstractDomain.java
package kpn.lib.domain;

public abstract class AbstractDomain<I> implements Domain<I> {
    private I id;

    @Override
    public void setId(I id) {
        this.id = id;
    }

    @Override
    public I getId() {
        return id;
    }

    @Override
    public String getInfo() {
        return "-";
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((id == null) ? 0 : id.hashCode());
        return result;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        AbstractDomain<I> other = (AbstractDomain<I>) obj;
        if (id == null) {
            if (other.id != null)
                return false;
        } else if (!id.equals(other.id))
            return false;
        return true;
    }
}